+++ clean.sh
@@ -25,9 +25,9 @@ for app in "${APPS[@]}"; do
     echo "Cleaning $app ..."
     local_app_dir="$APPS_DIR/$app"
     if [ -d "$local_app_dir" ]; then
-        rm -rf "$local_app_dir/bin"
-        rm -rf "$local_app_dir/obj"
-        rm -rf "$local_app_dir/perfdata"
+        rm -rf "${local_app_dir:?}/bin"
+        rm -rf "${local_app_dir:?}/obj"
+        rm -rf "${local_app_dir:?}/perfdata"
         rm -f "$local_app_dir"/*.binlog
     fi
     find "$BUILD_DIR/" -type d -name "${app}*" -exec rm -rf {} + 2>/dev/null
